+++ AngularClient/src/main/core/project-editor/project-editor.module.ts
@@ -1,6 +1,8 @@
 import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
 import { FormsModule } from "@angular/forms";
+import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
+import { BrowserModule } from "@angular/platform-browser";
 
 import { IProjectEditorConfig, ProjectEditorComponent } from "./project-editor.component";
 import { SymbolicPipe } from "./symbolic.pipe";
@@ -21,6 +23,8 @@ export { IProjectEditorConfig, NgChartModule }
   ],
   imports: [
     CommonModule,
+    BrowserAnimationsModule,
+    BrowserModule,
     FormsModule,
     NgChartModule,
     QRCodeModule,