+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **0.86.7**
 
 * Project Editor - Timer delay duration has been updated
+* Project Editor - Remove element item fix
 * "angularx-qrcode" package has been added - https://github.com/cordobo/angularx-qrcode
   * QRCodeJs library has been changed with "angularx-qrcode" library
   * (QRCodeJs library has been removed)