+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 * "angularx-qrcode" package has been added - https://github.com/cordobo/angularx-qrcode
   * QRCodeJs library has been changed with "angularx-qrcode" library
   * (QRCodeJs library has been removed)
-* ElementCell - setDecimalValue() instead of currentUserDecimalValue() and allUsersDecimalValue() was created
+* ElementCell -  instead of setDecimalValue(), currentUserDecimalValue() and allUsersDecimalValue() was created
+* ElementItem updated
 
 **0.86.6**